fix: merged IngestingHeader to IngestingDatasets component, and fetch data in frontend side.
